+++ src/main/java/racingcar/domain/car/CarName.java
@@ -1,4 +1,4 @@
-package racingcar.car;
+package racingcar.domain.car;
 
 public class CarName {
