+++ src/test/java/racingcar/domain/move/MoveTest.java
@@ -1,20 +1,21 @@
-package racingcar.move;
+package racingcar.domain.move;
 
 import org.junit.jupiter.api.Test;
-import racingcar.move.MoveNumber;
-import racingcar.move.MoveType;
+import racingcar.domain.move.Move;
+import racingcar.domain.move.MoveType;
+
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-public class MoveNumberTest {
+public class MoveTest {
 
     @Test
     void 주어진_값이_4이상이면_전진한다() {
-        MoveNumber moveNumber = MoveNumber.valueOf(4);
+        Move move = new Move(() -> 4);
 
-        MoveType moveType = moveNumber.get();
+        MoveType moveType = move.goOrStop();
 
         assertThat(moveType).isSameAs(MoveType.GO);
         assertThat(moveType).isNotSameAs(MoveType.STOP);
@@ -22,9 +23,9 @@ public class MoveNumberTest {
 
     @Test
     void 주어진_값이_3이하이면_멈춘다() {
-        MoveNumber moveNumber = MoveNumber.valueOf(3);
+        Move move = new Move(() -> 3);
 
-        MoveType moveType = moveNumber.get();
+        MoveType moveType = move.goOrStop();
 
         assertThat(moveType).isSameAs(MoveType.STOP);
         assertThat(moveType).isNotSameAs(MoveType.GO);
@@ -33,16 +34,20 @@ public class MoveNumberTest {
     @Test
     void 주어진_값이_1에서_9까지가_아니라면_에러발생() {
         assertDoesNotThrow(() -> {
-            MoveNumber.valueOf(9);
-            MoveNumber.valueOf(1);
+            Move move9 = new Move(() -> 9);
+            Move move1 = new Move(() -> 1);
+            move9.goOrStop();
+            move1.goOrStop();
         });
 
         IllegalArgumentException exception10 = assertThrows(IllegalArgumentException.class, () -> {
-            MoveNumber.valueOf(10);
+            Move move10 = new Move(() -> 10);
+            move10.goOrStop();
         });
 
         IllegalArgumentException exception0 = assertThrows(IllegalArgumentException.class, () -> {
-            MoveNumber.valueOf(0);
+            Move move0 = new Move(() -> 0);
+            move0.goOrStop();
         });
 
         assertThat(exception10.getMessage()).isEqualTo("숫자는 1에서 9까지만 입력 가능합니다.");
