+++ src/test/java/racingcar/domain/move/PositionTest.java
@@ -1,7 +1,8 @@
-package racingcar.move;
+package racingcar.domain.move;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.move.Position;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
